updated expr_extension.py and test code to add toggle/add/remove function
